refactored with better rails 3 generators, vendor assets and now structured to allow for orm substitution
