+++ lib/generators/reportable/assets/templates/flot/NOTES
@@ -1,0 +1,7 @@
+
+  ** Beware that flot depends on jQuery!
+  ** Don't forget to include excanvas.min.js, jquery.flot.min.js as well as jQuery in your layout's head section, e.g.
+  **
+  **  <%= javascript_include_tag 'excanvas.min', 'jquery.flot.min' %>
+  **
+
