+++ lib/saulabs/reportable/engine.rb
@@ -1,0 +1,7 @@
+module Saulabs
+	module Reportable
+		class Engine < ::Rails::Engine
+			# ensures vendor/assets of this gem are added to Rails app asset pipeline
+		end
+	end
+end
